+++ src/App.tsx
@@ -102,7 +102,11 @@ const App = () => {
         setDarkMode={setDarkMode}
         setViewModeGrouped={setViewModeGrouped}
       >
-        <SnackbarProvider maxSnack={2} autoHideDuration={5000}>
+        <SnackbarProvider
+          maxSnack={2}
+          autoHideDuration={5000}
+          anchorOrigin={{ horizontal: "center", vertical: "bottom" }}
+        >
           <ConfirmProvider>
             <Sync
               opened={openSync}